Update output files location

+++ src/net/buddat/wgenerator/WGenerator.java
@@ -486,7 +486,7 @@ public class WGenerator extends JFrame implements ActionListener, FocusListener 
 	public WurmAPI getAPI() {
 		if (api == null)
 			try {
-				api = WurmAPI.create("./", (int) (Math.log(heightMap.getMapSize()) / Math.log(2)));
+				api = WurmAPI.create("./maps/" + heightMap.getSeed() + "/", (int) (Math.log(heightMap.getMapSize()) / Math.log(2)));
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
@@ -859,9 +859,9 @@ public class WGenerator extends JFrame implements ActionListener, FocusListener 
 			
 			MapData map = getAPI().getMapData();
 			try {
-				ImageIO.write(map.createMapDump(), "png", new File("map.png"));
-				ImageIO.write(map.createTopographicDump(true, (short) 250), "png", new File("topography.png"));
-				ImageIO.write(map.createCaveDump(true), "png", new File("cave.png"));
+				ImageIO.write(map.createMapDump(), "png", new File("./maps/" + heightMap.getSeed() + "/map.png"));
+				ImageIO.write(map.createTopographicDump(true, (short) 250), "png", new File("./maps/" + heightMap.getSeed() + "/topography.png"));
+				ImageIO.write(map.createCaveDump(true), "png", new File("./maps/" + heightMap.getSeed() + "/cave.png"));
 			} catch (IOException ex) {
 				logger.log(Level.SEVERE, null, ex);
 			}
@@ -875,11 +875,20 @@ public class WGenerator extends JFrame implements ActionListener, FocusListener 
 			
 			updateAPIMap();
 			getAPI().getMapData().saveChanges();
+			getAPI().close();
+			
+			api = null;
+			updateAPIMap();
 		}
 		
 		if (e.getSource() == btnSaveActions) {
+			if (tileMap == null) {
+				JOptionPane.showMessageDialog(this, "TileMap does not exist - Add Dirt first", "Error Saving Map", JOptionPane.ERROR_MESSAGE);
+				return;
+			}
+			
 			try {
-				File actionsFile = new File("./map_actions.txt");
+				File actionsFile = new File("./maps/" + heightMap.getSeed() + "/map_actions.txt");
 				actionsFile.createNewFile();
 				
 				BufferedWriter bw = new BufferedWriter(new FileWriter(actionsFile));
@@ -893,8 +902,13 @@ public class WGenerator extends JFrame implements ActionListener, FocusListener 
 		}
 		
 		if (e.getSource() == btnLoadActions) {
+			if (tileMap == null) {
+				JOptionPane.showMessageDialog(this, "TileMap does not exist - Add Dirt first", "Error Saving Map", JOptionPane.ERROR_MESSAGE);
+				return;
+			}
+			
 			try {
-				File actionsFile = new File("./map_actions.txt");
+				File actionsFile = new File("./maps/" + heightMap.getSeed() + "/map_actions.txt");
 				
 				BufferedReader br = new BufferedReader(new FileReader(actionsFile));
 				String line;
